+++ src/esruntime.rs
@@ -15,6 +15,7 @@ use crate::spidermonkeyruntimewrapper::SmRuntime;
 use std::cell::RefCell;
 use std::time::Duration;
 
+use hirofa_utils::js_utils::Script;
 use hirofa_utils::task_manager::TaskManager;
 
 lazy_static! {
@@ -28,31 +29,11 @@ pub struct EsRuntime {
     inner: Arc<EsRuntimeInner>,
 }
 
-pub struct EsScriptCode {
-    absolute_path: String,
-    script_code: String,
-}
-
-impl EsScriptCode {
-    pub fn new(absolute_path: String, script_code: String) -> Self {
-        Self {
-            absolute_path,
-            script_code,
-        }
-    }
-    pub fn get_path(&self) -> &str {
-        self.absolute_path.as_str()
-    }
-    pub fn get_code(&self) -> &str {
-        self.script_code.as_str()
-    }
-}
-
 /// A ModuleCodeLoader function is used to load code into the runtime
 /// The first argument is the (relative) path of the module to import
 /// The second argument is the absolute path to the module which is importing the new module (reference_path)
 /// the EsScriptCode struct which is returned should allways contain an absolute path even if the module is loaded with a relative path
-pub type ModuleCodeLoader = dyn Fn(&str, &str) -> Option<EsScriptCode> + Send + Sync + 'static;
+pub type ModuleCodeLoader = dyn Fn(&str, &str) -> Option<Script> + Send + Sync + 'static;
 
 impl EsRuntime {
     /// create a builder to instantiate an EsRuntime
@@ -254,9 +235,10 @@ impl EsRuntime {
 #[cfg(test)]
 pub mod tests {
 
-    use crate::esruntime::{EsRuntime, EsScriptCode};
+    use crate::esruntime::EsRuntime;
     use crate::esvaluefacade::EsValueFacade;
     use crate::jsapi_utils::EsErrorInfo;
+    use hirofa_utils::js_utils::Script;
     use log::LevelFilter;
     use std::sync::Arc;
     use std::thread;
@@ -280,16 +262,10 @@ pub mod tests {
                     "i'm a little teapot short and stout, Tip me over and pour me out!, {}",
                     path
                 );
-                Some(EsScriptCode {
-                    absolute_path: path.to_string(),
-                    script_code: code,
-                })
+                Some(Script::new(path, code.as_str()))
             } else {
                 let code = format!("export default () => 123; export const other = Math.sqrt(8); console.log('running imported test module'); \n\nconsole.log('parsing a module from code loader for filename: {}');", path);
-                Some(EsScriptCode {
-                    absolute_path: path.to_string(),
-                    script_code: code,
-                })
+                Some(Script::new(path, code.as_str()))
             }
         };
         let rt = EsRuntime::builder()